`fn dav1d_sgr_filter1_neon`: Deduplicate w/ generics
